requirements.txt
pandas==1.1.0
joblib==0.16.0
numpy==1.19.1
matplotlib==3.3.0
xgboost==1.2.1
streamlit==0.64.0
seaborn==0.10.1
Pillow==8.0.1
scikit_learn==0.23.2
sweetviz==1.1.2
yellowbrick==1.2
